Consolidate remote_read etc. into `GetValues`
We have a need for fetching 'raw' datapoints. So instead of leaving this
out in just the servergroup path we are instead adding it to our API
interface. This means we can use it through our multi_api etc. to reduce
code complexity
