+++ promclient/interface.go
@@ -6,6 +6,7 @@ import (
 
 	v1 "github.com/prometheus/client_golang/api/prometheus/v1"
 	"github.com/prometheus/common/model"
+	"github.com/prometheus/prometheus/pkg/labels"
 )
 
 // API Subset of the interface defined in the prometheus client
@@ -18,4 +19,6 @@ type API interface {
 	QueryRange(ctx context.Context, query string, r v1.Range) (model.Value, error)
 	// Series finds series by label matchers.
 	Series(ctx context.Context, matches []string, startTime time.Time, endTime time.Time) ([]model.LabelSet, error)
+	// GetValue loads the raw data for a given set of matchers in the time range
+	GetValue(ctx context.Context, start, end time.Time, matchers []*labels.Matcher) (model.Value, error)
 }
